Move client link tests to tests/tracking/ for tracing SDK compatibility
The tracing SDK CI (core check) installs mlflow-tracing, which doesn't
include sqlparse or the model registry stack. MlflowClient imports
pull in the full tracking dependency chain, crashing test collection
for the entire file. Move the two client tests to
tests/tracking/test_span_links_client.py so they run in the full
MLflow CI instead.
